+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
 base initial commit
+- abstract class for store and action